tests/requirements.txt
tensorflow-cpu
scikit-image
scipy
